requirements.txt

streamlit==1.39.0
pdfplumber==0.11.4
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.10
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.0
huggingface-hub==0.24.5
tqdm==4.66.4
numpy==1.26.4
scikit-learn==1.4.2
faiss-cpu==1.8.0
